refactor(sandbox): host/worker 境界の明確化と PluginHostManager の責務分離
sandbox パッケージのファイルがフラットに並び、Host (main thread) と
Worker (guest) の境界、および「在籍管理 / 入力共有 / lifecycle」の責務が
1 つの巨大クラス (PluginHostManager 809行) に同居していたのを整理。

ディレクトリ構成:
  host/   … main thread 側 (Worker 管理 / DOM 描画 / 入力収集 / fetch / 診断)
  worker/ … Worker (guest) 側の隔離実行環境エントリ (sandbox.worker.ts)

PluginHostManager の責務分離:
  - PluginRegistry.ts  … 全 Worker の在籍簿 + emit ルーティング (repository 層)
                          static だった registry/routeEmit を module singleton 化
  - SharedInputPool.ts … 全 Worker で 1 つの InputCollector を共有 (infra 層)
                          参照カウント・カーソル・scroll element 配線
  - PluginHostManager.ts … 1 Worker の lifecycle のみ (usecase 層)
  - types.ts           … HostHandlers / PluginWorkerInfo / Options
  - capability.ts      … CAPABILITY_COMMANDS / CMD_TO_HANDLER 表

API 変更:
  - PluginHostManager.routeEmit (static) → routeEmit (named export) に変更。
    consumer は WorkerPluginHost の 1 箇所のみ更新。
  - registry / activeWorkerCount static getter は外部未使用のため
    getActiveWorkerCount() 等の関数に置換。

挙動への変更はなし。tsgo / biome / build:workers / vite build すべて通過
(sandbox.worker は dist/assets/sandbox.worker-*.js として正しく解決される)。
