+++ packages/sandbox/src/host/pluginDiagnostics.ts
